+++ client/src/Components/Header/Navbar/Goal.css
@@ -1,0 +1,11 @@
+.goal{
+    margin-right: 35%;
+    margin-top: 1%;
+    color: aliceblue;
+}
+.button{
+    background-color:var(--bg);
+    border-style: none;
+    color: azure;
+    
+}
