+++ client/src/Components/Header/Navbar/Navbar.js
@@ -8,6 +8,7 @@ import DropdownMenu from "../Navbar/DropdownMenu";
 import developmentSvg from "../../../Assets/svg/development.svg";
 
 import navOptions from "../../../Assets/svg/selectDropdown.svg";
+import Goal from "./Goal";
 
 
 
@@ -15,6 +16,7 @@ const Navbar = () => {
     return (
 			<nav className={classes.navbar}>
 				<ul className={classes.navbarNav}>
+					<Goal />
 					<NavItem navText="Dev Planner" href="/myPlan" icon={developmentSvg} />
 					<NavItem navText="Options" icon={navOptions}>
 						<DropdownMenu />
